fixed comment unit test and moved notifications component into the component folder
