+++ fakebook/src/app/app.module.ts
@@ -6,7 +6,7 @@ import { PostViewComponent } from './component/post-view/post-view.component';
 import { CommentFormComponent } from './component/comment-form/comment-form.component';
 import { LikeViewComponent } from './component/like-view/like-view.component';
 import { NewsfeedComponent } from './component/newsfeed/newsfeed.component';
-import { NotificationsViewComponent } from './notifications-view/notifications-view.component';
+import { NotificationsViewComponent } from './component/notifications-view/notifications-view.component';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { HubConnection, HubConnectionBuilder } from '@aspnet/signalr';
 import { TimeAgoPipe } from 'time-ago-pipe';
