+++ fakebook/src/app/component/comment-form/comment-form.component.spec.ts
@@ -5,6 +5,7 @@ import { of } from 'rxjs';
 import { Comment } from 'src/app/model/comment';
 import { AuthService } from 'src/app/service/auth.service';
 import { CommentService } from 'src/app/service/comment.service';
+import { NotificationsService } from 'src/app/service/notifications.service';
 
 import { CommentFormComponent } from './comment-form.component';
 
@@ -19,6 +20,10 @@ describe('CommentFormComponent', () => {
     getUser(): void { }
   };
 
+  const fakeNotifService = {
+    createCommentNotification(): void { }
+  }
+
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [ CommentFormComponent ],
@@ -26,7 +31,8 @@ describe('CommentFormComponent', () => {
         { provide: CommentService, useValue: fakeCommentService },
         { provide: ActivatedRoute, useValue: {
           paramMap: of( convertToParamMap( { userId: 1 } ) ) } },
-        { provide: OktaAuthService, useValue: fakeAuthService}
+        { provide: OktaAuthService, useValue: fakeAuthService},
+        { provide: NotificationsService, useValue: fakeNotifService}
       ]
     })
     .compileComponents();
@@ -38,6 +44,6 @@ describe('CommentFormComponent', () => {
 
     const fakeCommentServ = TestBed.inject(CommentService);
     const fakeAuthServiceTwo = TestBed.inject(OktaAuthService);
-    component = new CommentFormComponent(fakeCommentServ, TestBed.inject(ActivatedRoute), fakeAuthServiceTwo);
+    component = new CommentFormComponent(fakeCommentServ, TestBed.inject(ActivatedRoute), fakeAuthServiceTwo, TestBed.inject(NotificationsService));
   });
 });
