+++ fakebook/src/app/component/notifications-view/notifications-view.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
-import { NotificationsService } from '../service/notifications.service';
-import { Notification } from '../model/notification';
+import { NotificationsService } from '../../service/notifications.service';
+import { Notification } from '../../model/notification';
 import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { runInThisContext } from 'vm';
-import { AuthService } from '../service/auth.service';
+import { AuthService } from '../../service/auth.service';
 
 @Component({
   selector: 'app-notifications-view',
